# apps/backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/sanity": {
            "get": {
                "operationId": "SanityController_findByCharacter",
                "parameters": [
                    {
                        "name": "characterId",
                        "required": true,
                        "in": "query",
                        "description": "Character ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Get all sanity conditions for a character",
                "tags": [
                    "sanity"
                ]
            },
            "post": {
                "operationId": "SanityController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateSanityConditionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "summary": "Create a new sanity condition",
                "tags": [
                    "sanity"
                ]
            }
        },
        "/api/sanity/{id}": {
            "patch": {
                "operationId": "SanityController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Sanity condition ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateSanityConditionDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Update a sanity condition",
                "tags": [
                    "sanity"
                ]
            },
            "delete": {
                "operationId": "SanityController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Sanity condition ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Delete a sanity condition",
                "tags": [
                    "sanity"
                ]
            }
        },
...
    },
    "info": {
        "title": "Game-Plug API",
        "description": "Call of Cthulhu 7e RPG Platform API - Complete backend specification with NestJS and OpenAPI 3.0",
        "version": "1.0.0",
        "contact": {
            "name": "Game-Plug Team",
            "url": "https://github.com/robinswood",
            "email": "contact@game-plug.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
...
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Development"
        },
        {
            "url": "http://localhost:4000",
            "description": "Production"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateSanityConditionDto": {
                "type": "object",
                "properties": {
                    "characterId": {
                        "type": "string",
                        "description": "Character ID associated with this condition"
                    },
                    "type": {
                        "type": "string",
                        "description": "Type of sanity condition",
                        "enum": [
                            "phobia",
                            "mania",
                            "temporary_insanity",
                            "indefinite_insanity"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the sanity condition"
                    },
                    "description": {
                        "type": "string",
                        "description": "Detailed description of the condition"
                    },
                    "duration": {
                        "type": "string",
                        "description": "Duration type",
                        "enum": [
                            "temporary",
                            "indefinite",
                            "permanent"
                        ]
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Whether the condition is currently active"
                    }
                },
                "required": [
                    "characterId",
                    "type",
                    "name"
                ]
            },
            "UpdateSanityConditionDto": {
                "type": "object",
                "properties": {
                    "characterId": {
                        "type": "string",
                        "description": "Character ID associated with this condition"
                    },
                    "type": {
                        "type": "string",
                        "description": "Type of sanity condition",
                        "enum": [
                            "phobia",
                            "mania",
                            "temporary_insanity",
                            "indefinite_insanity"
                        ]
                    },
                    "name": {
                        "type": "string",
                        "description": "Name of the sanity condition"
                    },
                    "description": {
                        "type": "string",
                        "description": "Detailed description of the condition"
                    },
                    "duration": {
                        "type": "string",
                        "description": "Duration type",
                        "enum": [
                            "temporary",
                            "indefinite",
                            "permanent"
                        ]
                    },
                    "isActive": {
                        "type": "boolean",
                        "description": "Whether the condition is currently active"
                    }
                }
            },
...
        }
    }
}