# apps/backend/src/modules/sanity/dto/update-sanity-condition.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateSanityConditionDto } from './create-sanity-condition.dto';

export class UpdateSanityConditionDto extends PartialType(CreateSanityConditionDto) {}